+++ corehq/apps/app_manager/tests/test_shadow_forms.py
@@ -2,7 +2,10 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 from django.test import SimpleTestCase
 
-from corehq.apps.app_manager.models import AdvancedOpenCaseAction, LoadUpdateAction, AdvancedFormActions
+from corehq.apps.app_manager.models import (
+    AdvancedOpenCaseAction,
+    LoadUpdateAction,
+)
 from corehq.apps.app_manager.tests.app_factory import AppFactory
 from corehq.apps.app_manager.tests.util import TestXmlMixin, extract_xml_partial
